% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{BDT.feedback.no_score}
\alias{BDT.feedback.no_score}
\title{BDT feedback (no score)}
\usage{
BDT.feedback.no_score(dict = BDT::BDT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
BDT_demo(feedback = BDT.feedback.no_score())}
}
