% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{BDT.feedback.simple_score}
\alias{BDT.feedback.simple_score}
\title{BDT feedback (simple score)}
\usage{
BDT.feedback.simple_score(dict = BDT::BDT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant's score is reported at the end of the test.
}
\examples{
\dontrun{
standalone_BDT(feedback = BDT.feedback.simple_score())}
}
