% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BDT_standalone}
\alias{BDT_standalone}
\title{Standalone BDT}
\usage{
BDT_standalone(
  title = "Beat Drop Alignment Test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL,
  languages = BDT_languages(),
  dict = BDT::BDT_dict,
  feedback = NULL,
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), French (\code{"French"}),
and German (\code{"DE"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{feedback}{(Null or a page)
if NULL ther last page will show the "You completed the test" message along with "Your results have been save", else
it will only show the latter but the feedback specified here.
Currently, BDT.feedback.no_score() and BDT.feedback.simple_score() are supported (s. feedback.R).
You can also design your own feedback function.}

\item{...}{Further arguments to be passed to \code{\link{BDT}()}.}
}
\description{
This function launches a standalone testing session for the BDT.
This can be used for data collection, either in the laboratory or online.
}
